AzureFunctionsProject/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

asgiref==3.7.2
asttokens==2.4.0
azure-common==1.1.28
azure-core==1.29.3
azure-eventhub==5.11.4
azure-functions==1.17.0
azure-identity==1.14.0
azure-keyvault-secrets==4.7.0
backcall==0.2.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
comm==0.1.4
cryptography==41.0.3
DateTime==5.2
debugpy==1.6.7.post1
decorator==5.1.1
Django==4.2.5
django-stubs==4.2.4
django-stubs-ext==4.2.2
exceptiongroup==1.1.3
executing==1.2.0
idna==3.4
ipykernel==6.25.2
ipython==8.15.0
isodate==0.6.1
jedi==0.19.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jwt==1.3.1
matplotlib-inline==0.1.6
msal==1.23.0
msal-extensions==1.0.0
mypy==1.5.1
mypy-extensions==1.0.0
nest-asyncio==1.5.7
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.10.0
portalocker==2.7.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
tornado==6.3.3
traitlets==5.9.0
types-pytz==2023.3.0.1
types-PyYAML==6.0.12.11
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
websocket-client==1.6.2
zope.interface==6.0
